chore: Add delete functionality to EditDebt component
This commit adds the ability to delete a debt in the EditDebt component. It includes a "Delete Debt" button that triggers a confirmation modal. When the user confirms the deletion, the debt is deleted from the database and the user is redirected to the home page. This feature enhances the application by allowing users to easily remove debts they no longer need.